--- specification/_types/termvectors/TermVector.ts
@@ -1,47 +1,0 @@
-/*
- * Licensed to Elasticsearch B.V. under one or more contributor
- * license agreements. See the NOTICE file distributed with
- * this work for additional information regarding copyright
- * ownership. Elasticsearch B.V. licenses this file to you under
- * the Apache License, Version 2.0 (the "License"); you may
- * not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *    http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing,
- * software distributed under the License is distributed on an
- * "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
- * KIND, either express or implied.  See the License for the
- * specific language governing permissions and limitations
- * under the License.
- */
-
-import { Dictionary } from '@spec_utils/Dictionary'
-import { double, integer, long } from '@_types/Numeric'
-
-export class TermVector {
-  field_statistics: FieldStatistics
-  terms: Dictionary<string, Term>
-}
-
-export class FieldStatistics {
-  doc_count: integer
-  sum_doc_freq: long
-  sum_ttf: long
-}
-
-export class Term {
-  doc_freq?: integer
-  score?: double
-  term_freq: integer
-  tokens: Token[]
-  ttf?: integer
-}
-
-export class Token {
-  end_offset?: integer
-  payload?: string
-  position: integer
-  start_offset?: integer
-}
